@@ -21,7 +21,7 @@
     ],
     "type": "library",
     "require": {
-        "php": ">=7.2.0"
+        "php": "~8.1"
     },
     "require-dev": {
         "mockery/mockery": "^1.3",
